slack bot

+++ slackbot/swirhentv/plugins/mention.py
@@ -14,7 +14,9 @@ slack = Slacker(slackbot_settings.API_TOKEN)
 def torrent_search(message, argment):
     message.send('さがすー')
     resultfile='temp/tss.result'
-    os.remove(resultfile)
+    if os.path.exists(resultfile):
+        os.remove(resultfile)
+
     cmd = './tss.sh {0}'.format(argment)
     call_cmd(cmd)
     if os.path.exists(resultfile):
